Add test on <identifier> type attribute
Added a test on the type attribute of the <identifier> to only
process the <identifier> elements that have the type attribute
set to "local". This is needed because the CLIO info, when
present,is also stored in an <identifier> element - in this case,
however, the <identifier> will not have a type attribute,
according to the CUL specs.