Do not add linefeed at end of output, new implementation
Before, the code was using two templates, with the mode attribute
set in one to differentiate, to handle the special case of the last
line of the output. However, this entailed duplicating a bit of code.
This new implementation uses <xsl:with-param> and <xsl:param> to pass
in the end-of-line character.